pyproject.toml
[project]
name = "pcomp-experiments"
version = "0.1.0"
description = ""
readme = "README.md"
authors = [{ name = "Cameron Pitsch", email = "cameron.pitsch@rwth-aachen.de" }]
requires-python = ">=3.10"
dependencies = [
    "mpire>=2.10.0",
    "nbformat>=4.2.0",
    "numpy<2.0",
    "pcomp",
    "plotly>=6.0.1",
    "pydantic>=2.9.2",
    "pyyaml>=6.0.2",
    "seaborn>=0.13.2",
    "streamlit>=1.38",
]

[dependency-groups]
dev = [
    "ipykernel>=6.29.5",
    "mypy>=1.15.0",
    "pandas-stubs>=2.2.3.250308",
    "plotly-stubs>=0.0.5",
    "types-pyyaml>=6.0.12.20250402",
    "types-seaborn>=0.13.2.20250111",
]

[tool.uv.sources]
pcomp = { path = "../pcomp" }
